FundraisingDjango/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.7.4.post0=py38h96a0964_0
amqp=5.0.6=pypi_0
asgiref=3.4.1=pyhd3eb1b0_0
async-timeout=3.0.1=py38hecd8cb5_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pyhd3eb1b0_0
base58=2.1.0=pyhd8ed1ab_0
billiard=3.6.4.0=pypi_0
bitarray=1.2.2=py38haf1e3a3_0
brotlipy=0.7.0=py38h9ed2024_1003
ca-certificates=2021.7.5=hecd8cb5_1
celery=5.1.2=pypi_0
certifi=2021.5.30=py38hecd8cb5_0
cffi=1.14.6=py38h2125817_0
chardet=4.0.0=py38hecd8cb5_1003
click=7.1.2=pypi_0
click-didyoumean=0.0.3=pypi_0
click-plugins=1.1.1=pypi_0
click-repl=0.2.0=pypi_0
crispy-bootstrap5=0.4=pypi_0
cryptography=3.4.7=py38h2fd3fbb_0
cytoolz=0.11.0=py38haf1e3a3_0
django=3.2.5=pyhd3eb1b0_0
django-celery-beat=2.2.0=pypi_0
django-compat=1.0.15=pypi_0
django-crispy-forms=1.12.0=pypi_0
django-crontab=0.7.1=pypi_0
django-mathfilters=1.0.0=pypi_0
django-timezone-field=4.2.1=pypi_0
djangorestframework=3.12.4=pypi_0
eth-abi=2.1.1=py_2
eth-account=0.5.4=py_0
eth-hash=0.3.1=pyhd8ed1ab_0
eth-keyfile=0.5.1=pyh9f0ad1d_1
eth-keys=0.3.3=pyh9f0ad1d_0
eth-rlp=0.2.1=py_0
eth-typing=2.2.2=py_0
eth-utils=1.10.0=py38h50d1736_0
gmp=6.2.1=h23ab428_2
h11=0.12.0=pyhd3eb1b0_0
h2=3.2.0=py38_1
hexbytes=0.2.1=py_0
hpack=3.0.0=py_0
httpcore=0.13.3=pyhd3eb1b0_0
httpx=0.18.2=pyhd3eb1b0_0
hyperframe=5.2.0=py_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
ipfshttpclient=0.7.0=pyhd8ed1ab_0
jsonschema=3.2.0=py_2
kombu=5.1.0=pypi_0
krb5=1.19.2=hcd88c3b_0
libcxx=10.0.0=1
libedit=3.1.20210216=h9ed2024_1
libffi=3.3=hb1e8313_2
libpq=12.2=h1b4eb34_1
libprotobuf=3.17.2=h2842e9f_1
lru-dict=1.1.7=py38h96a0964_0
multiaddr=0.0.9=py_0
multidict=5.1.0=py38h9ed2024_2
mysql-connector-c=6.1.11=hccea1a4_1
mysqlclient=2.0.3=py38h23ab428_1
ncurses=6.2=h0a44026_1
netaddr=0.7.19=py_0
openssl=1.1.1k=h9ed2024_0
parsimonious=0.8.1=py_0
pip=21.1.3=py38hecd8cb5_0
prompt-toolkit=3.0.19=pypi_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.17.2=py38h23ab428_0
psycopg2=2.8.6=py38hbcfaee0_1
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pycryptodome=3.10.1=py38h659dea6_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyrsistent=0.17.3=py38haf1e3a3_0
pysocks=1.7.1=py38_1
python=3.8.10=h88f2d9e_7
python-crontab=2.5.1=pypi_0
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytz=2021.1=pyhd3eb1b0_0
readline=8.1=h9ed2024_0
redis=3.5.3=pypi_0
requests=2.25.1=pyhd3eb1b0_0
rfc3986=1.4.0=py_0
rlp=1.2.0=pyh9f0ad1d_0
setuptools=52.0.0=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=py38hecd8cb5_1
sqlite=3.36.0=hce871da_0
sqlparse=0.4.1=py_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.1=pyhd3eb1b0_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
varint=1.0.2=pyh9f0ad1d_0
vine=5.0.0=pypi_0
wcwidth=0.2.5=py_0
web3=5.21.0=py38h50d1736_0
websockets=9.1=py38h96a0964_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.6.3=py38h9ed2024_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
